move sliver app bar tests to correct file
